auth_validator/requirements.txt
fastapi
uvicorn
PyJWT
python-dotenv
cryptography